Neural network

+++ WifiMeasure/src/menthallab/wifimeasure/Instance.java
@@ -4,19 +4,19 @@ import java.util.*;
 
 public class Instance
 {
-	private final Map<String, Integer> values;
+	private final Map<String, Double> values;
 	
 	public Instance()
 	{
-		this.values = new HashMap<String, Integer>();
+		this.values = new HashMap<String, Double>();
 	}
 	
-	public void add(String attributeName, int value)
+	public void add(String attributeName, double value)
 	{
 		this.values.put(attributeName, value);
 	}
 	
-	public Integer get(String attributeName)
+	public Double get(String attributeName)
 	{
 		return this.values.get(attributeName);
 	}
